requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
eventlet>=0.30.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
gevent>=22.10.2
gevent-websocket>=0.10.1
google-auth==2.38.0
google-genai==0.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
json_repair==0.35.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0a1
pydantic_core==2.28.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
pyzmq==26.2.1
ratelimit==2.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarg==0.1.9
